cnn_predictor/requirements.txt
torch
torchvision
pillow
scikit-learn
matplotlib